Waybar script now checks for files and commands

+++ scripts/waybar-reloadable.sh
@@ -1,7 +1,37 @@
 #!/usr/bin/env bash
 
+if [[ ! -d "$HOME/.config/waybar" ]]; then
+	echo -e "\\nERROR: Missing waybar folder. Make sure you have your config and style loaded."
+	exit 1
+fi
+
+if [[ ! -f "$HOME/.config/waybar/config" ]]; then
+	echo -e "\\nERROR: Missing waybar config."
+	exit 1
+fi
+
+if [[ ! -f "$HOME/.config/waybar/style.css" ]]; then
+	echo -e "\\nERROR: Missing waybar CSS stylesheet."
+	exit 1
+fi
+
 # Add any waybar files so they can be watched
 CONFIG_FILES="$HOME/dotfiles/config/waybar/config $HOME/dotfiles/config/waybar/style.css"
+
+if ! command -v trap > /dev/null ; then
+	echo -e "\\nERROR: \"trap\" command not found"
+	exit 1
+fi
+
+if ! command -v inotifywait > /dev/null ; then
+	echo -e "\\nERROR: \"inotify-tools\" not found"
+	exit 1
+fi
+
+if ! command -v killall > /dev/null ; then
+	echo -e "\\nERROR: \"killall\" not found"
+	exit 1
+fi
 
 # If this script stops, kill waybar
 trap "killall waybar" EXIT
